refactor: flatten source tree into single compiler/ directory
Previously, source files were organized across compiler/src/,
compiler/src/Sema/, and compiler/include/ as separate subdirectories.
This commit collapses the entire compiler source tree directly into the
compiler/ root directory, eliminating the src/ and include/ split.

Changes:
- Remove `compiler/include/*.h` (`ast.h`, `codegen.h`, `common.h`, `error.h`,
  all stupid header files)
- Moved `compiler/src/*.c` to root `compiler/`
- Add flat equivalents directly under compiler/:
    - urusc.c      (formerly src/main.c)
    - urusc.h      (new unified header, replaces separate include/*.h)
    - urusctok.h   (token definitions, extracted from common header)
    - misc.c       (formerly src/util.c, renamed for clarity)
    - sema.c       (Sema/scope.c and Sema/scope.h)
    - builtins.c   (formerly src/Sema/builtins.c, now top-level)
    - lexer.c, ast.c, parser.c, codegen.c, error.c, preprocess.c
      (relocated from src/, functionality unchanged)
- Update CMakeLists.txt:
    - Adjust configure_file path from include/config.h.in to config.h.in
    - Update SOURCES list to use flat paths
    - Remove target_include_directories for the old include/ directory (cmakeconfig)

Motivation: the include/ + src/ split added unnecessary indirection for
a single-binary compiler of this scale. A flat layout is simpler to
navigate, easier to grep, and removes the friction of maintaining
mirrored .h/.c pairs across two directories.

No functional changes to compiler behavior.
