--- compiler/include/common.h
@@ -1,45 +1,0 @@
-/*
- * Copyright 2026 Urus Foundation (https://github.com/Urus-Foundation)
- *
- * This file is part of the Urus Programming Language.
- * For more about this language check at
- *
- *    https://github.com/Urus-Foundatation/Urus
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#ifndef URUS_COMMON_H
-#define URUS_COMMON_H
-
-#include <limits.h>
-#include <stdio.h>
-#include <stdlib.h>
-
-// --  Path Separator  --
-#ifdef _WIN32
-#define URUSC_PATHSEP '\\'
-#else
-#define URUSC_PATHSEP '/'
-#endif
-
-// --  Path Maximum  --
-#ifndef PATH_MAX
-#ifdef MAX_PATH
-#define PATH_MAX MAX_PATH
-#else
-#define PATH_MAX 4096 /* safe default value */
-#endif
-#endif
-
-#endif
