+++ compiler/parser.c
@@ -19,17 +19,7 @@
  * limitations under the License.
  */
 
-#ifndef _WIN32
-#define _POSIX_C_SOURCE 200809L
-#endif
-
-#include "parser.h"
-#include "error.h"
-#include "util.h"
-#include <errno.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
+#include "urusc.h"
 
 // ---- Helpers ----
